rename Legend to RasterStyle
